requirements.txt
fastapi
uvicorn
python-dotenv
llama-index
openai